requirements.txt
djangorestframework-jwt
djangorestframework
Django~=2.2.4